+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. The version
 
 ### Added
 * Support for not having engines in all directions.
+* Settings for the yaw/pitch/roll controller.
 
 ### Removed
 * Fuel gauges on the screen element have been removed to save code space (still available on the HUD).